Builder: implement builder for directory and stargz
Introduce a new Builder trait, it accepts build context parameter,
then builds source into bootstrap(for directory and stargz source)
and blob(only for directory source). we can also implement Builder
for diff build source in the future.
